List the online git repositories of the current user

List the online git repositories of the current user

Usage

list_github_repositories(username = get_current_git_username)

Arguments

username

string. The git username to list the repositories. Default is the current git username.

Value

A list of the current public repositories for the current git user

Examples

list_github_repositories("cran")
#> GitHub repositories for user cran :
#> -  https://github.com/cran/A3 
#> -  https://github.com/cran/AalenJohansen 
#> -  https://github.com/cran/aaMI 
#> -  https://github.com/cran/aaSEA 
#> -  https://github.com/cran/AATtools 
#> -  https://github.com/cran/aba 
#> -  https://github.com/cran/ABACUS 
#> -  https://github.com/cran/abasequence 
#> -  https://github.com/cran/abbreviate 
#> -  https://github.com/cran/abbyyR 
#> -  https://github.com/cran/abc 
#> -  https://github.com/cran/abc.data 
#> -  https://github.com/cran/ABC.RAP 
#> -  https://github.com/cran/abcADM 
#> -  https://github.com/cran/ABCanalysis 
#> -  https://github.com/cran/abcdeFBA 
#> -  https://github.com/cran/ABCExtremes 
#> -  https://github.com/cran/abclass 
#> -  https://github.com/cran/ABCoptim 
#> -  https://github.com/cran/ABCp2 
#> -  https://github.com/cran/abcrf 
#> -  https://github.com/cran/abcrlda 
#> -  https://github.com/cran/abctools 
#> -  https://github.com/cran/abd 
#> -  https://github.com/cran/abdiv 
#> -  https://github.com/cran/abe 
#> -  https://github.com/cran/abemus 
#> -  https://github.com/cran/aberrance 
#> -  https://github.com/cran/abess 
#> -  https://github.com/cran/abf2 
if (FALSE) { # \dontrun{
list_github_repositories()
} # }