Start the import process from GitHub or your local json preference files

Import preferences by cloning a link or pulling from your GitHub. The function will start to pull from the currently active GitHub repository, then will import the local .json preference files. The function will look for the specific .json files only, in the given file path. If cloning is chosen, the folder will be inside the working directory.

Usage

start_import_prefs(
  preference_path = "R/rstudio_preferences/",
  pull_github = TRUE,
  clone_git = FALSE,
  git_url = NULL,
  git_path = NULL
)

Arguments

preference_path

string. Input the file path where your preferences json files are located.

pull_github

logical. Use TRUE to pull or clone from GitHub, otherwise FALSE to just import your local files.

clone_git

logical. Default is FALSE. Clone git for first time usage. A convenience function to avoid doing it manually, but it's essentially the same as doing gert::git_clone("github_url"). It will clone the repository inside your current working directory.

git_url

string. Allows to user the bypass the request of the git url to directly clone the desired git url repository to the working directory.

git_path

string. The path to clone the git repository. Default is the current working directory.

Value

Cloned repository and/or preference files copied to RStudio's preference folder

Details

Function wrapper of import_from_github() followed by import_local_prefs()

See also

Examples