Start the import process from Github by default or your local json preference files

Import preferences by cloning a link or pulling from your GitHub. The function will start to pull from the currently active GitHub repository, then will import the local .json preference files. The function will look for the specific .json files only, in the given file path. If cloning is chosen, the folder will be inside the working directory.

Usage

start_import_prefs(
  preference_path = "R/rstudio_preferences/",
  pull_github = TRUE,
  clone_git = FALSE,
  git_url = NULL
)

Arguments

preference_path

string. Input the file path where your preferences.json files are located.

pull_github

boolean. Use TRUE to pull or clone from GitHub, otherwise FALSE to just import your local files.

clone_git

boolean. Default is FALSE. Clone git for first time usage. A convenience function to avoid doing it manually, but it's essentially the same as doing gert::git_clone("github_url"). It will clone the repository inside your current working directory.

git_url

string. Allows to user the bypass the request of the git url to directly clone the desired git url repository to the working directory.

Value

Clone repository or preference files copied to RStudio's preference folder

Details

Function wrapper of import_from_github() followed by import_local_prefs()

See also

Examples

# To import your preferences from locally stored preferences files:

# Dummy preference files
temp_dir <- tempdir()
fs::file_create(file.path(temp_dir, "addins.json"))
fs::file_create(file.path(temp_dir, "rstudio_bindings.json"))
fs::file_create(file.path(temp_dir, "r.snippets"))
fs::file_create(file.path(temp_dir, "rstudio-prefs.json"))
start_import_prefs(pull_github = FALSE, preference_path = temp_dir)
#> ✔ Sucessfully imported RStudio preferences.
#> ✔ addins.json, r.snippets, and rstudio_bindings.json copied to
#> /home/runner/.config/rstudio/keybindings
#> ✔ rstudio-prefs.json copied to /home/runner/.config/rstudio.
#> Error in .rs.restartR(): could not find function ".rs.restartR"

# To import your preferences from Github:
try(start_import_prefs())
#> Error : Failed to evaluate glue component {gert::git_find(git_path)}
#> Caused by error:
#> ! failed to resolve path 'R/rstudio_preferences/': No such file or directory